BUG/MINOR: stream: Don't release a stream if FLT_END is still registered
When at least one filter is registered on a stream, the FLT_END analyzer is
called on both direction when all other analyzers have finished their
processing. During this step, filters may release any allocated elements if
necessary. So it is important to not skip it.

Unfortunately, if both stream interfaces are closed, it is possible to not wait
the end of this analyzer. It is possible to be in this situation if a filter
must wait and prevents the analyzer completion. To fix the bug, we now wait
FLT_END analyzer is no longer registered on both direction before releasing the
stream.

This patch may be backported as far as 1.7, but AFAIK, no filter is affected by
this bug. So the backport seems to be optional for now. In any case, it should
remain under observation for some weeks first.