MINOR: lua: Add a function to get a filter attached to a channel class
For now, there is no support for filters written in lua. So this function, if
called, will always return NULL. But when it will be called in a filter context,
it will return the filter structure attached to a channel class. This function
is also responsible to set the offset of data that may be processed and the
length of these data. When called outside a filter context (so from an action),
the offset is the input data position and the length is the input data
length. From a filter, the offset and the length of data that may be filtered
are retrieved the filter context.