MEDIUM: lua: Process buffer data using an offset and a length
The main change is that following functions will now process channel's data
using an offset and a length:

 * hlua_channel_dup_yield()
 * hlua_channel_get_yield()
 * hlua_channel_getline_yield()
 * hlua_channel_append_yield()
 * hlua_channel_set()
 * hlua_channel_send_yield()
 * hlua_channel_forward_yield()

So for now, the offset is always the input data position and the length is the
input data length. But with the support for filters, from a filter context,
these values will be relative to the filter.

To make all processing clearer, the function _hlua_channel_dup() has been
updated and _hlua_channel_dupline(), _hlua_channel_append() and
_hlua_channel_erase_data() have been added.

This patch is mandatory to allow the support of the filters written in lua.