MEDIUM: filters/lua: Be prepared to filter TCP payloads
For filters written in lua, the tcp payloads will be filtered using methods
exposed by the Channel class. So the corrsponding C binding functions must be
prepared to process payload in a filter context and not only in an action
context. Following functions are concerned: hlua_channel_dup_yield(),
hlua_channel_get_yield(), hlua_channel_getline_yield(),
hlua_channel_append_yield(), hlua_channel_set_yield(), hlua_channel_send_yield()
and hlua_channel_forward_yield().

The main change is the offset where to start to process data in the
channel buffer, and the length of these data. For an action, all input data are
considered. But for a filter, it depends on what the filter is allow to forward
when the tcp_payload callback function is called. It depends on previous calls
but also on other filters.

There is also another difference when these functions are called from a
filter. In an action context, these functions may yield. But it is not possible
in a filter context. Some filter callback functions may yield but not those
filtering the payload.

Finally, when one of these functions is called from a lua filter, a test is
performed to be sure it is during the tcp payload filtering.

Note that for now, it is not possible to define callbacks to filter data. The
documentation has not been updated for now.