MEDIUM: filters/lua: Support declaration of some filter callback functions in lua
It is now possible to write some filter callback functions in lua. All filter
callbacks are not supported yet but the mechanism to call them is now in
place. Following method may be defined in the Lua filter class to be bound on
filter callbacks:

  * Filter:start_analyse(txn, chn)
  * Filter:end_analyse(txn, chn)
  * Filter:tcp_payload(txn, chn, offset, length)

hlua_filter_callback() function is responsible to call the good lua function
depending on the filter callback function. Using some flags it is possible to
allow a lua call to yield or no, to retrieve a return value or not, and to
specify if a channel or an http message must be passed as second argument. For
now, the HTTP part has not been added yet. It is also possible to add extra
argument adding them on the stack before the call.

3 new functions are exposed by the global object "filter". The first one,
filter.wake_time(ms_delay), to set the wake_time when a Lua callback function
yields (if allowed). The two others, filter.register_data_filter(flt, chn) and
filter.unregister_data_filter(flt, chn), to enable or disable the data filtering
on a channel for a specific lua filter instance.

Here is a simple example :

    MyFilter = {}
    MyFilter.id = "My Lua filter"
    MyFilter.flags = filter.FLT_CFG_FL_HTX
    MyFilter.__index = MyFilter

    function MyFilter:new()
        flt = {}
        setmetatable(flt, MyFilter)
        flt.req_len = 0
        flt.res_len = 0
        return flt
     end

    function MyFilter:start_analyze(txn, chn)
        filter.register_data_filter(flt, chn)
    end

    function MyFilter:end_analyze(txn, chn)
        print("<Total> request: "..flt.req_len.." - response: "..flt.res_len)
    end

    function MyFilter:tcp_payload(txn, chn, offset, len)
        if chn:is_resp() then
            flt.res_len = flt.res_len + len
            print("<TCP:Response> offset: "..offset.." - length: "..len)
        else
            flt.req_len = flt.req_len + len
            print("<TCP:Request> offset: "..offset.." - length: "..len)
        end
    end