WIP/BUG/MAJOR: filters/lua: Fix how lua references are created/duplicated
Lua filters are not working at all since the last rebase. When a script is
loaded per-thread, the references on the filter objects are mixed between
threads. Especially when the same filter is used several times.

So, everything should be fixed now. This patch is a temporary patch only. It
should be merged with preceding ones. But for now, I prefer to keep it
separate to track changes. It is still very experimental, and it is helpfull
to keep the history for now.